+++ src-backend/tauri/src/adapter/exporter/avatar_explorer/mod.rs
@@ -1,5 +1,0 @@
-mod builder;
-mod definitions;
-mod logic;
-
-pub use logic::export_as_avatar_explorer_compatible_structure;