+++ src/pages/Preference/hook.ts
@@ -1,11 +1,7 @@
 import { useState } from 'react'
 
 export type PreferenceTabIDs =
-  | 'settings'
-  | 'adapter'
-  | 'statistics'
-  | 'logs'
-  | 'about'
+  'settings' | 'adapter' | 'statistics' | 'logs' | 'about'
 
 type ReturnProps = {
   activeTab: PreferenceTabIDs