+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ### 🚀新機能
 
 ### 🐛バグ修正
+- アセット追加ボタンを高速で連打したときに二重登録される問題を修正
 
 ### ⚙️その他
+- Avatar Explorer への移行サポートを廃止
+  - [最新の Avatar Explorer](https://github.com/puk06/VRC-Avatar-Explorer/releases/latest) に KonoAsset からデータを移行する機能があるため、そちらをご利用ください
 
 ## 1.2.6